+++ pkg/clientwrappers/dataflow/dataflow.go
@@ -1,8 +1,9 @@
 package dataflow
 
 import (
-	"github.com/Orange-OpenSource/nifikop/pkg/util/clientconfig"
 	"strings"
+
+	"github.com/Orange-OpenSource/nifikop/pkg/util/clientconfig"
 
 	"github.com/Orange-OpenSource/nifikop/api/v1alpha1"
 	"github.com/Orange-OpenSource/nifikop/pkg/clientwrappers"
@@ -156,7 +157,7 @@ func IsOutOfSyncDataflow(
 
 	return isParameterContextChanged(parameterContext, processGroups) ||
 		isVersioningChanged(flow, registry, pGEntity) || !isVersionSync(flow, pGEntity) || localChanged(pGEntity) ||
-		isParentProcessGroupChanged(flow, config, pGEntity) || isNameChanged(flow, pGEntity), nil
+		isParentProcessGroupChanged(flow, config, pGEntity) || isNameChanged(flow, pGEntity) || isPostionChanged(flow, pGEntity), nil
 }
 
 func isParameterContextChanged(
@@ -209,6 +210,13 @@ func isVersioningChanged(
 		flow.Spec.FlowId != pgFlowEntity.Component.VersionControlInformation.FlowId ||
 		flow.Spec.BucketId != pgFlowEntity.Component.VersionControlInformation.BucketId ||
 		registry.Status.Id != pgFlowEntity.Component.VersionControlInformation.RegistryId
+}
+
+// isPostionChanged check if the position of the process group is out of sync.
+func isPostionChanged(flow *v1alpha1.NifiDataflow, pgFlowEntity *nigoapi.ProcessGroupEntity) bool {
+	return flow.Spec.FlowPosition != nil &&
+		((flow.Spec.FlowPosition.X != nil && float64(*flow.Spec.FlowPosition.X) != pgFlowEntity.Component.Position.X) ||
+			(flow.Spec.FlowPosition.Y != nil && float64(*flow.Spec.FlowPosition.Y) != pgFlowEntity.Component.Position.Y))
 }
 
 // SyncDataflow implements the logic to sync a NifiDataflow with the deployed flow.
@@ -264,9 +272,27 @@ func SyncDataflow(
 		return RemoveDataflow(flow, config)
 	}
 
-	if isNameChanged(flow, pGEntity) {
+	if isNameChanged(flow, pGEntity) || isPostionChanged(flow, pGEntity) {
 		pGEntity.Component.ParentGroupId = flow.Spec.GetParentProcessGroupID(config.RootProcessGroupId)
 		pGEntity.Component.Name = flow.Name
+
+		var xPos, yPos float64
+		if flow.Spec.FlowPosition == nil || flow.Spec.FlowPosition.X == nil {
+			xPos = pGEntity.Component.Position.X
+		} else {
+			xPos = float64(*flow.Spec.FlowPosition.X)
+		}
+
+		if flow.Spec.FlowPosition == nil || flow.Spec.FlowPosition.Y == nil {
+			yPos = pGEntity.Component.Position.Y
+		} else {
+			yPos = float64(*flow.Spec.FlowPosition.Y)
+		}
+
+		pGEntity.Component.Position = &nigoapi.PositionDto{
+			X: xPos,
+			Y: yPos,
+		}
 		_, err := nClient.UpdateProcessGroup(*pGEntity)
 		if err := clientwrappers.ErrorUpdateOperation(log, err, "Stop flow"); err != nil {
 			return nil, err
@@ -698,16 +724,34 @@ func updateProcessGroupEntity(
 	}
 
 	if entity.Component == nil {
-		entity.Component = &nigoapi.ProcessGroupDto{
-			Position: &nigoapi.PositionDto{
-				X: 1,
-				Y: 1,
-			},
-		}
+		entity.Component = &nigoapi.ProcessGroupDto{}
 	}
 
 	entity.Component.Name = flow.Name
 	entity.Component.ParentGroupId = flow.Spec.GetParentProcessGroupID(config.RootProcessGroupId)
+
+	var xPos, yPos float64
+	if entity.Component.Position != nil {
+		xPos = entity.Component.Position.X
+		yPos = entity.Component.Position.Y
+	} else {
+		if flow.Spec.FlowPosition == nil || flow.Spec.FlowPosition.X == nil {
+			xPos = float64(1)
+		} else {
+			xPos = float64(*flow.Spec.FlowPosition.X)
+		}
+
+		if flow.Spec.FlowPosition == nil || flow.Spec.FlowPosition.Y == nil {
+			yPos = float64(1)
+		} else {
+			yPos = float64(*flow.Spec.FlowPosition.Y)
+		}
+	}
+
+	entity.Component.Position = &nigoapi.PositionDto{
+		X: xPos,
+		Y: yPos,
+	}
 	entity.Component.VersionControlInformation = &nigoapi.VersionControlInformationDto{
 		GroupId:          stringFactory(),
 		RegistryName:     stringFactory(),